% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectGameUrls.R
\name{collect_game_urls}
\alias{collect_game_urls}
\title{Collect game urls for a competition by year/season}
\usage{
collect_game_urls(
  competitions,
  years = NULL,
  from = NULL,
  to = NULL,
  season = FALSE
)
}
\arguments{
\item{competitions}{a vector of competitions to search, check \code{rugby_competitions} tibble for accepted values}

\item{years}{a vector of years ; cannot be used with from and to}

\item{from}{a year to start searching from}

\item{to}{a year to search until ; from and to must be used together}

\item{season}{if set to \code{TRUE} the competition is assumed to span multiple years, usually for Northern hemisphere competitions}
}
\value{
returns a list of rugbypass.com game urls
}
\description{
Extract game urls from a rugbypass competition landing page.
These urls identify an individual game and end in ".../stats/"
}
