% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaData.R
\name{get_game_metadata}
\alias{get_game_metadata}
\title{Parse a rugbypass.com url to retrieve team level stats}
\usage{
get_game_metadata(data, is_html = FALSE)
}
\arguments{
\item{data}{a rugbypass.com url to retrieve data from. Can aslo be previously extracted html}

\item{is_html}{is the data a url or extracted html. Default is `FALSE`` i.e. the data variable holds a url}
}
\value{
Returns tibble with containing game meta data (e.g. data, venue, score) from rugbypass.com
}
\description{
Extract the team statistics from a rugbypassmatch url.
These urls identify an individual game and end in ".../stats/"
}
