% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerStats.R
\name{get_player_stats}
\alias{get_player_stats}
\title{Parse a rugbypass.com url to retrieve player level stats}
\usage{
get_player_stats(data, is_html = FALSE)
}
\arguments{
\item{data}{a rugbypass.com url to retrieve data from. Can aslo be previously extracted html}

\item{is_html}{is the data a url or extracted html. Default is \code{FALSE} i.e. the data variable holds a url}
}
\value{
Returns tibble with containing player level statistics from rugbypass.com
}
\description{
Extract the player statistics from a rugbypass match url.
These urls identify an individual game and end in ".../stats/"
}
