% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_meta_data.R
\docType{data}
\name{rugby_stadiums}
\alias{rugby_stadiums}
\title{Parse a rugbypass.com url to retrieve team level stats}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 175 rows and 5 columns.
}
\usage{
rugby_stadiums
}
\arguments{
\item{data}{a rugbypass.com url to retrieve data from. Can aslo be previously extracted html}

\item{is_html}{is the data a url or extracted html. Default is `FALSE`` i.e. the data variable holds a url}

\item{wide}{return the data in wide or long format. Deafult is `TRUE` for wide data}
}
\value{
Returns tibble with containing player level statistics from rugbypass.com
}
\description{
Extract the team statistics from a rugbypassmatch url.
These urls identify an individual game and end in ".../stats/"
}
\keyword{datasets}
