% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_rugbypass.R
\name{scrape_rugbypass}
\alias{scrape_rugbypass}
\title{Scrape a rugbypass.com url}
\usage{
scrape_rugbypass(url, as_list = FALSE)
}
\arguments{
\item{as_list}{return data in 3 element list: \code{meta_data}, \code{team_stats} and \code{player_stats}; Default is \code{FALSE}}

\item{urls}{rugbypass.com url to retrieve data from.}
}
\value{
Returns nested tibble (or list) containing game, team and player level statistics from rugbypass.com
}
\description{
Extract data from a rugbypass match url.
These urls identify an individual game and end in ".../stats/"
}
