% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_meta_data.R
\name{get_game_metadata}
\alias{get_game_metadata}
\title{Parse a rugbypass.com url to retrieve team level stats}
\usage{
get_game_metadata(url)
}
\arguments{
\item{url}{rugbypass.com url to retrieve data from. Can aslo be previously extracted html}
}
\value{
Returns tibble with containing player level statistics from rugbypass.com
}
\description{
Extract the team statistics from a rugbypassmatch url.
These urls identify an individual game and end in ".../stats/"
}
